Refactor: 마일리지 타입 변경

+++ src/main/java/com/skhu/moodfriend/app/dto/member/resDto/MemberInfoResDto.java
@@ -7,7 +7,7 @@ import lombok.Builder;
 public record MemberInfoResDto(
         String email,
         String name,
-        long mileage,
+        int mileage,
         String loginType
 ) {
     public static MemberInfoResDto of(Member member) {
